requirements.txt
openpyxl==3.1.5
pandas==3.0.0
Pillow==12.1.0
xlrd==2.0.2
xlutils==2.0.0
xlwt==1.3.0
